Move other utils out of ArrayUtils
